try and extract api setup/teardown to separate file
